requirements/base.txt
# Core
pytz==2022.7.1
Django==4.1.7
django-configurations==2.4
gunicorn==20.1.0
newrelic==8.7.0

# For the persistence stores
psycopg2-binary==2.9.5
dj-database-url==1.2.0

# Model Tools
django-model-utils==4.3.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.14.0
django-filter==22.1

# Static and Media Storage
django-storages==1.13.2
boto3==1.26.85
Pillow==9.4.0

# Testing
mock==5.0.1
factory-boy==3.2.1
django-nose==1.4.7
nose-progressive==1.5.2
coverage==7.2.1


# Utils
setuptools==58